remove pnpm-lock and use package-lock.json instead.
Pnpm has a great stats and version UI and great performance. But it heavily assumes we used semantic versioning, which was absolutely not the case in pre release 5.0. Causing really difficult to debug issues with packages being used from cache instead of downloaded when sub dependencies are only "patch updates"... for feathers, thats 37 versions of 5.0 beta with many many breaking changes.